+++ R/percentVJ.R
@@ -14,7 +14,9 @@
 #' @param input.data The product of \code{\link{combineTCR}}, 
 #' \code{\link{combineBCR}}, or \code{\link{combineExpression}}.
 #' @param chain "TRA", "TRB", "TRG", "TRG", "IGH", "IGL"
-#' @param group.by The variable to use for grouping.
+#' @param group.by The variable to use for grouping
+#' @param order.by A vector of specific plotting order or "alphanumeric"
+#' to plot groups in order
 #' @param exportTable Returns the data frame used for forming the graph
 #' @param palette Colors to use in visualization - input any 
 #' \link[grDevices]{hcl.pals}.
@@ -28,6 +30,7 @@
 percentVJ <- function(input.data,
                       chain = "TRB",
                       group.by = NULL, 
+                      order.by = NULL,
                       exportTable = FALSE, 
                       palette = "inferno") {
   
@@ -86,6 +89,12 @@ percentVJ <- function(input.data,
   
   #Melting matrix and Visualizing
   mat_melt <- melt(mat)
+  if(!is.null(order.by)) {
+    mat_melt <- .ordering.function(vector = order.by,
+                                   group.by = "L1", 
+                                   mat_melt)
+  }
+  
   plot <- ggplot(mat_melt, aes(y=Var1, x = Var2, fill=round(value*100,2))) +
     geom_tile(lwd= 0.25, color = "black") +
     scale_fill_gradientn(name = "Percentage", colors = .colorizer(palette,21)) +