+++ R/vizGenes.R
@@ -104,6 +104,8 @@ vizGenes <- function(input.data,
     col.lab <- "Scaled Values"
     values <- "proportion"
   }
+  #Ensure only x.axis match is displayed
+  input.data <- input.data[grep(x.axis, input.data[,x.axis]),]
   
   #Calculating the summary values
   if (!is.null(y.axis) && y.axis != "element.names") {
@@ -193,7 +195,7 @@ vizGenes <- function(input.data,
               "TRBD" = 2, "TRDD" = 2, "IGHV" = 2, 
               "TRBJ" = 3, "TRDJ" = 2, "IGHJ" = 3)
   chain.poisiton <- chains[position]
-  if(substring(position,3,3) %in% c("A", "G", "K", "L")) {
+  if(substring(position,3,3) %in% c("A", "G", "H")) {
     chain.gene <- str_split(df[,"CTgene"], "_", simplify = TRUE)[,1]
   } else {
     chain.gene <- str_split(df[,"CTgene"], "_", simplify = TRUE)[,2]