+++ tests/testthat/test-percentAA.R
@@ -6,11 +6,20 @@ test_that("percentAA works", {
 
   set.seed(42)
   expect_doppelganger(
+    "percentAA_order_plot",
+    percentAA(combined, 
+              chain = "TRB", 
+              aa.length = 20,
+              order.by = c("P17B","P18B","P19B","P20B","P17L","P18L","P19L","P20L"))
+  )
+  
+  expect_doppelganger(
     "percentAA_plot",
     percentAA(combined, 
               chain = "TRB", 
               aa.length = 20)
   )
+  
   set.seed(42)
   expect_equal(
     percentAA(combined,