+++ tests/testthat/test-percentVJ.R
@@ -10,6 +10,14 @@ test_that("percentVJ works", {
     percentVJ(combined, 
               chain = "TRB")
   )
+  
+  expect_doppelganger(
+    "percentVJ_order_plot",
+    percentVJ(combined, 
+              chain = "TRB",
+              order.by = c("P17B","P18B","P19B","P20B","P17L","P18L","P19L","P20L"))
+  )
+  
   set.seed(42)
   expect_equal(
     percentVJ(combined,